Removed org test that's not relevant anymore (create w/o name or email)
Validation was recently added at the hackathon to require name and email
email address when creating organizations; this is in direct conflict
with the test being removed and I have no idea why it was added or is
relevant as it's not associated with a commit that needs it.  Will find
out later when working on import logic if a real requirement.