Event's source is not always defined in from_abstract_event
Support was added for assigning an organization to an event from the
event's source, but some tests do not include a source.  All known
scenarios in ActivateHub for converting abstract events to events do
provide a source; however, rather than changing the tests and increase
incompatibilities with upstream calagator at this point, it makes sense
to assume source is optional.