Remove :dependent => :destroy for assocations that should persist
It's way too dangerous to cascade deletions when sources are destroyed;
it removes all the events and venues from the calendar, which is unexpected
behavior and and makes rolling back really painful.

Also removing :dependent => :nullify, as it does not trigger any callbacks
and thus create any entries in PaperTrail to allow rolling back.  The code
base appears unaffected by the change.  WHEN we have support for rolling
back changes in bulk, and across multiple models, we might look at
restoring behavior similar to :dependent => :nullify.